OLTU-196 - Type cast error in JWTClaimsSetParser
applied patch submitted by Amrit Verma


git-svn-id: https://svn.apache.org/repos/asf/oltu/trunk/oauth-2.0/jwt@1749341 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/oltu/oauth2/jwt/io/JWTClaimsSetParser.java
@@ -34,9 +34,9 @@ final class JWTClaimsSetParser extends CustomizableEntityReader<JWT, JWT.Builder
         if (AUDIENCE.equals(key)) {
             handleAudience(value);
         } else if (EXPIRATION_TIME.equals(key)) {
-            getBuilder().setClaimsSetExpirationTime(((Integer) value).longValue());
+            getBuilder().setClaimsSetExpirationTime(((Long) value));
         } else if (ISSUED_AT.equals(key)) {
-            getBuilder().setClaimsSetIssuedAt(((Integer) value).longValue());
+            getBuilder().setClaimsSetIssuedAt(((Long) value));
         } else if (ISSUER.equals(key)) {
             getBuilder().setClaimsSetIssuer(String.valueOf(value));
         } else if (JWT_ID.equals(key)) {
